Fixed warnings

+++ source/renderer/DX12/Helpers/DXResource.hpp
@@ -21,7 +21,7 @@ public:
 
     CD3DX12_RESOURCE_DESC GetDesc() const { return mDesc; }
     D3D12_RESOURCE_STATES GetState() const { return mState; }
-    size_t GetResourceSize() const { return mResourceSize; }
+    uint32_t GetResourceSize() const { return mResourceSize; }
 
     void ChangeState(D3D12_RESOURCE_STATES dstState);
     bool mResizeBuffer = false;
@@ -30,6 +30,6 @@ private:
     D3D12_RESOURCE_STATES mState = D3D12_RESOURCE_STATE_COMMON;
     CD3DX12_RESOURCE_DESC mDesc{};
     ComPtr<ID3D12Resource> mResource;
-    size_t mResourceSize = 0;
+    uint32_t mResourceSize = 0;
     // std::vector<std::unique_ptr<DXResource>> mUploadBuffers;
 };
